% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imf.R
\name{imf_countries}
\alias{imf_countries}
\title{List IMF country codes}
\usage{
imf_countries()
}
\value{
tibble: iso (character), label (character)
}
\description{
Returns ~241 countries and regions with ISO codes.
}
